Export functions to another lua file
