+++ frontend/ui/data/creoptions.lua
@@ -1,11 +1,8 @@
 local Device = require("device")
-local InfoMessage = require("ui/widget/infomessage")
+local OptionsUtil = require("ui/data/optionsutil")
 local S = require("ui/data/strings")
-local UIManager = require("ui/uimanager")
-local Screen = Device.screen
-local T = require("ffi/util").template
-
 local _ = require("gettext")
+local Screen = Device.screen
 
 -- add multiply operator to Aa dict
 local Aa = setmetatable({"Aa"}, {
@@ -18,123 +15,12 @@ local Aa = setmetatable({"Aa"}, {
     end
 })
 
-local function enable_if_equals(configurable, option, value)
-    return configurable[option] == value
+local showValues = function(configurable, option, prefix)
+    OptionsUtil:showValues(configurable, option, prefix)
 end
 
-local function showValues(configurable, option)
-    local default = G_reader_settings:readSetting("copt_"..option.name)
-    local current = configurable[option.name]
-    local value_default, value_current
-    local suffix = option.name_text_suffix or ""
-    if option.name == "screen_mode" then
-        current = Screen:getScreenMode()
-    end
-    local arg_table = {}
-    if option.toggle and option.values then
-        for i=1,#option.toggle do
-            arg_table[option.values[i]] = option.toggle[i]
-        end
-    end
-    if not default then
-        default = "not set"
-        if option.toggle and option.values then
-            value_current = current
-            current = arg_table[current]
-        end
-    elseif option.toggle and option.values then
-        value_current = current
-        value_default = default
-        default = arg_table[default]
-        current = arg_table[current]
-    end
-    if option.labels and option.values then
-        for i=1,#option.labels do
-            if default == option.values[i] then
-                default = option.labels[i]
-                break
-            end
-        end
-        for i=1,#option.labels do
-            if current == option.values[i] then
-                current = option.labels[i]
-                break
-            end
-        end
-    end
-    if option.name_text_true_values and option.toggle and option.values and value_default then
-        UIManager:show(InfoMessage:new{
-            text = T(_("%1:\nCurrent value: %2 (%5%4)\nDefault value: %3 (%6%4)"), option.name_text,
-                current, default, suffix, value_current, value_default)
-        })
-    elseif option.name_text_true_values and option.toggle and option.values and not value_default then
-        UIManager:show(InfoMessage:new{
-            text = T(_("%1:\nCurrent value: %2 (%5%4)\nDefault value: %3"), option.name_text,
-                current, default, suffix, value_current)
-        })
-    else
-        UIManager:show(InfoMessage:new{
-            text = T(_("%1:\nCurrent value: %2%4\nDefault value: %3%4"), option.name_text, current,
-                default, suffix)
-        })
-    end
-end
-
-local function tableComp(a,b)
-    if #a ~= #b then return false end
-    for i=1,#a do
-        if a[i] ~= b[i] then return false end
-    end
-    return true
-end
-
-local function showValuesMargins(configurable, option)
-    local default = G_reader_settings:readSetting("copt_"..option.name)
-    local current = configurable[option.name]
-    local current_string
-    for i=1,#option.toggle do
-        if tableComp(current, option.values[i]) then
-            current_string = option.toggle[i]
-            break
-        end
-    end
-    if not default then
-        UIManager:show(InfoMessage:new{
-            text = T(_([[
-%1:
-Current value: %2
-  left: %3
-  top: %4
-  right: %5
-  bottom: %6
-Default value: not set]]),
-                option.name_text, current_string, current[1], current[2], current[3], current[4])
-        })
-    else
-        local default_string
-        for i=1,#option.toggle do
-            if tableComp(default, option.values[i]) then
-                default_string = option.toggle[i]
-                break
-            end
-        end
-        UIManager:show(InfoMessage:new{
-            text = T(_([[
-%1:
-Current value: %2
-  left: %3
-  top: %4
-  right: %5
-  bottom: %6
-Default value: %7
-  left: %8
-  top: %9
-  right: %10
-  bottom: %11]]),
-                option.name_text, current_string, current[1], current[2], current[3], current[4],
-                default_string, default[1], default[2], default[3], default[4])
-        })
-    end
+local showValuesMargins = function(configurable, option, prefix)
+    OptionsUtil:showValuesMargins(configurable, option, prefix)
 end
 
 local CreOptions = {
@@ -232,13 +118,13 @@ local CreOptions = {
                 event = "ChangeSize",
                 args = {"decrease", "increase"},
                 alternate = false,
-                name_text_hold_callback = function(configurable)
+                name_text_hold_callback = function(configurable, __, prefix)
                     local opt = {
                         name = "font_size",
                         name_text = _("Font Size"),
                     }
-                    showValues(configurable, opt)
-                end
+                    OptionsUtil:showValues(configurable, opt, prefix)
+                end,
             }
         }
     },
@@ -315,7 +201,7 @@ local CreOptions = {
                 default_arg = nil,
                 event = "ToggleEmbeddedFonts",
                 enabled_func = function(configurable)
-                    return enable_if_equals(configurable, "embedded_css", 1)
+                    return OptionsUtil:enableIfEquals(configurable, "embedded_css", 1)
                 end,
                 name_text_hold_callback = showValues,
             },
