+++ frontend/ui/data/koptoptions.lua
@@ -1,71 +1,11 @@
 local Device = require("device")
-local InfoMessage = require("ui/widget/infomessage")
+local OptionsUtil = require("ui/data/optionsutil")
 local S = require("ui/data/strings")
-local UIManager = require("ui/uimanager")
 local _ = require("gettext")
 local Screen = Device.screen
-local T = require("ffi/util").template
 
-local function enable_if_equals(configurable, option, value)
-    return configurable[option] == value
-end
-
-local function showValues(configurable, option)
-    local default = G_reader_settings:readSetting("kopt_"..option.name)
-    local current = configurable[option.name]
-    local value_default, value_current
-    local suffix = option.name_text_suffix or ""
-    if option.name == "screen_mode" then
-        current = Screen:getScreenMode()
-    end
-    local arg_table = {}
-    if option.toggle and option.values then
-        for i=1,#option.toggle do
-            arg_table[option.values[i]] = option.toggle[i]
-        end
-    end
-    if not default then
-        default = "not set"
-        if option.toggle and option.values then
-            value_current = current
-            current = arg_table[current]
-        end
-    elseif option.toggle and option.values then
-        value_current = current
-        value_default = default
-        default = arg_table[default]
-        current = arg_table[current]
-    end
-    if option.labels and option.values then
-        for i=1,#option.labels do
-            if default == option.values[i] then
-                default = option.labels[i]
-                break
-            end
-        end
-        for i=1,#option.labels do
-            if current == option.values[i] then
-                current = option.labels[i]
-                break
-            end
-        end
-    end
-    if option.name_text_true_values and option.toggle and option.values and value_default then
-        UIManager:show(InfoMessage:new{
-            text = T(_("%1:\nCurrent value: %2 (%5%4)\nDefault value: %3 (%6%4)"), option.name_text,
-                current, default, suffix, value_current, value_default)
-        })
-    elseif option.name_text_true_values and option.toggle and option.values and not value_default then
-        UIManager:show(InfoMessage:new{
-            text = T(_("%1:\nCurrent value: %2 (%5%4)\nDefault value: %3"), option.name_text,
-                current, default, suffix, value_current)
-        })
-    else
-        UIManager:show(InfoMessage:new{
-            text = T(_("%1:\nCurrent value: %2%4\nDefault value: %3%4"), option.name_text, current,
-                default, suffix)
-        })
-    end
+local showValues = function(configurable, option, prefix)
+    OptionsUtil:showValues(configurable, option, prefix)
 end
 
 local KoptOptions = {
@@ -99,7 +39,7 @@ local KoptOptions = {
                 enabled_func = Device.isTouchDevice,
                 event = "PageCrop",
                 args = {"manual", "auto", "semi-auto"},
-                name_text_hold_callback = showValues
+                name_text_hold_callback = showValues,
             }
         }
     },
@@ -114,7 +54,7 @@ local KoptOptions = {
                 default_value = DSCROLL_MODE,
                 event = "SetScrollMode",
                 args = {true, false},
-                name_text_hold_callback = showValues
+                name_text_hold_callback = showValues,
             },
             {
                 name = "full_screen",
@@ -125,7 +65,7 @@ local KoptOptions = {
                 event = "SetFullScreen",
                 args = {true, false},
                 show = false,
-                name_text_hold_callback = showValues
+                name_text_hold_callback = showValues,
             },
             {
                 name = "page_margin",
@@ -134,7 +74,7 @@ local KoptOptions = {
                 values = {0.05, 0.10, 0.25},
                 default_value = DKOPTREADER_CONFIG_PAGE_MARGIN,
                 event = "MarginUpdate",
-                name_text_hold_callback = showValues
+                name_text_hold_callback = showValues,
             },
             {
                 name = "line_spacing",
@@ -143,7 +83,7 @@ local KoptOptions = {
                 values = {1.0, 1.2, 1.4},
                 default_value = DKOPTREADER_CONFIG_LINE_SPACING,
                 advanced = true,
-                name_text_hold_callback = showValues
+                name_text_hold_callback = showValues,
             },
             {
                 name = "max_columns",
@@ -156,9 +96,9 @@ local KoptOptions = {
                 values = {1,2,3},
                 default_value = DKOPTREADER_CONFIG_MAX_COLUMNS,
                 enabled_func = function(configurable)
-                    return enable_if_equals(configurable, "text_wrap", 1)
+                    return OptionsUtil:enableIfEquals(configurable, "text_wrap", 1)
                 end,
-                name_text_hold_callback = showValues
+                name_text_hold_callback = showValues,
             },
             {
                 name = "justification",
@@ -174,10 +114,10 @@ local KoptOptions = {
                 default_value = DKOPTREADER_CONFIG_JUSTIFICATION,
                 advanced = true,
                 enabled_func = function(configurable)
-                    return enable_if_equals(configurable, "text_wrap", 1)
+                    return OptionsUtil:enableIfEquals(configurable, "text_wrap", 1)
                 end,
                 labels = {S.AUTO, S.LEFT, S.CENTER, S.RIGHT, S.JUSTIFY},
-                name_text_hold_callback = showValues
+                name_text_hold_callback = showValues,
             },
         }
     },
@@ -195,7 +135,7 @@ local KoptOptions = {
                 default_value = DKOPTREADER_CONFIG_FONT_SIZE,
                 event = "FontSizeUpdate",
                 enabled_func = function(configurable)
-                    return enable_if_equals(configurable, "text_wrap", 1)
+                    return OptionsUtil:enableIfEquals(configurable, "text_wrap", 1)
                 end,
             },
             {
@@ -209,14 +149,14 @@ local KoptOptions = {
                 args = {-0.05, 0.05},
                 alternate = false,
                 enabled_func = function(configurable)
-                    return enable_if_equals(configurable, "text_wrap", 1)
+                    return OptionsUtil:enableIfEquals(configurable, "text_wrap", 1)
                 end,
-                name_text_hold_callback = function(configurable)
+                name_text_hold_callback = function(configurable, __, prefix)
                     local opt = {
                         name = "font_size",
                         name_text = _("Font Size"),
                     }
-                    showValues(configurable, opt)
+                    showValues(configurable, opt, prefix)
                 end
             }
         }
@@ -234,7 +174,7 @@ local KoptOptions = {
                 event = "GammaUpdate",
                 args = {0.8, 1.0, 1.5, 2.0, 3.0, 4.0, 6.0, 9.0},
                 labels = {0.8, 1.0, 1.5, 2.0, 3.0, 4.0, 6.0, 9.0},
-                name_text_hold_callback = showValues
+                name_text_hold_callback = showValues,
             }
         }
     },
@@ -258,7 +198,7 @@ local KoptOptions = {
                         event = "InitScrollPageStates",
                     },
                 },
-                name_text_hold_callback = showValues
+                name_text_hold_callback = showValues,
             },
             {
                 name = "page_opt",
@@ -266,7 +206,7 @@ local KoptOptions = {
                 toggle = {S.ON, S.OFF},
                 values = {1, 0},
                 default_value = 0,
-                name_text_hold_callback = showValues
+                name_text_hold_callback = showValues,
             },
             {
                 name="doc_language",
@@ -276,7 +216,7 @@ local KoptOptions = {
                 default_value = DKOPTREADER_CONFIG_DOC_DEFAULT_LANG_CODE,
                 event = "DocLangUpdate",
                 args = DKOPTREADER_CONFIG_DOC_LANGS_CODE,
-                name_text_hold_callback = showValues
+                name_text_hold_callback = showValues,
             },
             {
                 name = "word_spacing",
@@ -285,9 +225,9 @@ local KoptOptions = {
                 values = DKOPTREADER_CONFIG_WORD_SPACINGS,
                 default_value = DKOPTREADER_CONFIG_DEFAULT_WORD_SPACING,
                 enabled_func = function(configurable)
-                    return enable_if_equals(configurable, "text_wrap", 1)
+                    return OptionsUtil:enableIfEquals(configurable, "text_wrap", 1)
                 end,
-                name_text_hold_callback = showValues
+                name_text_hold_callback = showValues,
             },
             {
                 name = "writing_direction",
@@ -296,9 +236,9 @@ local KoptOptions = {
                 values = {0, 1, 2},
                 default_value = 0,
                 enabled_func = function(configurable)
-                    return enable_if_equals(configurable, "text_wrap", 1)
+                    return OptionsUtil:enableIfEquals(configurable, "text_wrap", 1)
                 end,
-                name_text_hold_callback = showValues
+                name_text_hold_callback = showValues,
             },
             {
                 name = "quality",
@@ -308,9 +248,9 @@ local KoptOptions = {
                 default_value = DKOPTREADER_CONFIG_RENDER_QUALITY,
                 advanced = true,
                 enabled_func = function(configurable)
-                    return enable_if_equals(configurable, "text_wrap", 1)
+                    return OptionsUtil:enableIfEquals(configurable, "text_wrap", 1)
                 end,
-                name_text_hold_callback = showValues
+                name_text_hold_callback = showValues,
             },
             {
                 name = "forced_ocr",
@@ -319,7 +259,7 @@ local KoptOptions = {
                 values = {1, 0},
                 default_value = 0,
                 advanced = true,
-                name_text_hold_callback = showValues
+                name_text_hold_callback = showValues,
             },
             {
                 name = "defect_size",
@@ -330,9 +270,9 @@ local KoptOptions = {
                 event = "DefectSizeUpdate",
                 show = false,
                 enabled_func = function(configurable)
-                    return enable_if_equals(configurable, "text_wrap", 1)
+                    return OptionsUtil:enableIfEquals(configurable, "text_wrap", 1)
                 end,
-                name_text_hold_callback = showValues
+                name_text_hold_callback = showValues,
             },
             {
                 name = "auto_straighten",
@@ -342,9 +282,9 @@ local KoptOptions = {
                 default_value = DKOPTREADER_CONFIG_AUTO_STRAIGHTEN,
                 show = false,
                 enabled_func = function(configurable)
-                    return enable_if_equals(configurable, "text_wrap", 1)
+                    return OptionsUtil:enableIfEquals(configurable, "text_wrap", 1)
                 end,
-                name_text_hold_callback = showValues
+                name_text_hold_callback = showValues,
             },
             {
                 name = "detect_indent",
@@ -354,9 +294,9 @@ local KoptOptions = {
                 default_value = DKOPTREADER_CONFIG_DETECT_INDENT,
                 show = false,
                 enabled_func = function(configurable)
-                    return enable_if_equals(configurable, "text_wrap", 1)
+                    return OptionsUtil:enableIfEquals(configurable, "text_wrap", 1)
                 end,
-                name_text_hold_callback = showValues
+                name_text_hold_callback = showValues,
             },
         }
     },
