+++ frontend/ui/widget/configdialog.lua
@@ -235,7 +235,8 @@ function ConfigOption:init()
                     text_font_bold = false,
                     hold_callback = function()
                         if self.options[c].name_text_hold_callback then
-                            self.options[c].name_text_hold_callback(self.config.configurable, self.options[c])
+                            self.options[c].name_text_hold_callback(self.config.configurable, self.options[c],
+                                self.config.config_options.prefix)
                         end
                     end,
                 }
